Struct ssh2::Agent

pub struct Agent<'sess> { /* fields omitted */ }

A structure representing a connection to an SSH agent.

Agents can be used to authenticate a session.

Methods

impl<'sess> Agent<'sess>

pub fn connect(&mut self) -> Result<(), Error>

Connect to an ssh-agent running on the system.

pub fn disconnect(&mut self) -> Result<(), Error>

Close a connection to an ssh-agent.

pub fn list_identities(&mut self) -> Result<(), Error>

Request an ssh-agent to list of public keys, and stores them in the internal collection of the handle.

Call identities to get the public keys.

Important traits for Identities<'agent>
impl<'agent> Iterator for Identities<'agent> type Item = Result<PublicKey<'agent>, Error>;

pub fn identities(&self) -> Identities

Get an iterator over the identities of this agent.

pub fn userauth(
&self,
username: &str,
identity: &PublicKey
) -> Result<(), Error>

Attempt public key authentication with the help of ssh-agent.