Struct ssh2::Channel

pub struct Channel { /* fields omitted */ }

A channel represents a portion of an SSH connection on which data can be read and written.

Channels denote all of SCP uploads and downloads, shell sessions, remote process executions, and other general-purpose sessions. Each channel implements the Reader and Writer traits to send and receive data. Whether or not I/O operations are blocking is mandated by the blocking flag on a channel's corresponding Session.

Methods

impl Channel

pub fn setenv(&mut self, var: &str, val: &str) -> Result<(), Error>

Set an environment variable in the remote channel's process space.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

pub fn request_pty(
&mut self,
term: &str,
mode: Option<&str>,
dim: Option<(u32, u32, u32, u32)>
) -> Result<(), Error>

Request a PTY on an established channel.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

The dimensions argument is a tuple of (width, height, width_px, height_px)

pub fn request_pty_size(
&mut self,
width: u32,
height: u32,
width_px: Option<u32>,
height_px: Option<u32>
) -> Result<(), Error>

Request a PTY of a specified size

pub fn exec(&mut self, command: &str) -> Result<(), Error>

Execute a command

An execution is one of the standard process services defined by the SSH2 protocol.

Example

let mut channel = session.channel_session().unwrap();
channel.exec("ls").unwrap();
let mut s = String::new();
channel.read_to_string(&mut s).unwrap();
println!("{}", s);

pub fn shell(&mut self) -> Result<(), Error>

Start a shell

A shell is one of the standard process services defined by the SSH2 protocol.

pub fn subsystem(&mut self, system: &str) -> Result<(), Error>

Request a subsystem be started.

A subsystem is one of the standard process services defined by the SSH2 protocol.

pub fn process_startup(
&mut self,
request: &str,
message: Option<&str>
) -> Result<(), Error>

Initiate a request on a session type channel.

The SSH2 protocol currently defines shell, exec, and subsystem as standard process services.

Important traits for Stream<'channel>
impl<'channel> Read for Stream<'channel>impl<'channel> Write for Stream<'channel>

pub fn stderr<'a>(&'a mut self) -> Stream<'a>

Get a handle to the stderr stream of this channel.

The returned handle implements the Read and Write traits.

Important traits for Stream<'channel>
impl<'channel> Read for Stream<'channel>impl<'channel> Write for Stream<'channel>

pub fn stream<'a>(&'a mut self, stream_id: i32) -> Stream<'a>

Get a handle to a particular stream for this channel.

The returned handle implements the Read and Write traits.

Groups of substreams may be flushed by passing on of the following constants and then calling flush().

FLUSH_EXTENDED_DATA - Flush all extended data substreams

FLUSH_ALL - Flush all substreams

pub fn handle_extended_data(&mut self, mode: ExtendedData) -> Result<(), Error>

Change how extended data (such as stderr) is handled

pub fn exit_status(&self) -> Result<i32, Error>

Returns the exit code raised by the process running on the remote host at the other end of the named channel.

Note that the exit status may not be available if the remote end has not yet set its status to closed.

pub fn exit_signal(&self) -> Result<ExitSignal, Error>

Get the remote exit signal.

pub fn read_window(&self) -> ReadWindow

Check the status of the read window.

pub fn write_window(&self) -> WriteWindow

Check the status of the write window.

pub fn adjust_receive_window(
&mut self,
adjust: u64,
force: bool
) -> Result<u64, Error>

Adjust the receive window for a channel by adjustment bytes.

If the amount to be adjusted is less than the minimum adjustment and force is false, the adjustment amount will be queued for a later packet.

This function returns the new size of the receive window (as understood by remote end) on success.

pub fn eof(&self) -> bool

Check if the remote host has sent an EOF status for the channel. Take care: the EOF status is for the entire channel which can be confusing because the reading from the channel reads only the stdout stream. unread, buffered, stderr data will cause eof() to return false.

pub fn send_eof(&mut self) -> Result<(), Error>

Tell the remote host that no further data will be sent on the specified channel.

Processes typically interpret this as a closed stdin descriptor.

pub fn wait_eof(&mut self) -> Result<(), Error>

Wait for the remote end to send EOF. Note that unread buffered stdout and stderr will cause this function to return Ok(()) without waiting. You should call the eof() function after calling this to check the status of the channel.

pub fn close(&mut self) -> Result<(), Error>

Close an active data channel.

In practice this means sending an SSH_MSG_CLOSE packet to the remote host which serves as instruction that no further data will be sent to it. The remote host may still send data back until it sends its own close message in response.

To wait for the remote end to close its connection as well, follow this command with wait_closed

pub fn wait_close(&mut self) -> Result<(), Error>

Enter a temporary blocking state until the remote host closes the named channel.

Typically sent after close in order to examine the exit status.