Struct ssh2::Agent

pub struct Agent<'sess> {
    // some fields omitted
}

A structure representing a connection to an SSH agent.

Agents can be used to authenticate a session.

Methods

impl<'sess> Agent<'sess>

unsafe fn from_raw(sess: &Session, raw: *mut LIBSSH2_AGENT) -> Agent

Wraps a raw pointer in a new Agent structure tied to the lifetime of the given session.

This consumes ownership of raw.

fn connect(&mut self) -> Result<(), Error>

Connect to an ssh-agent running on the system.

fn disconnect(&mut self) -> Result<(), Error>

Close a connection to an ssh-agent.

fn list_identities(&mut self) -> Result<(), Error>

Request an ssh-agent to list of public keys, and stores them in the internal collection of the handle.

Call identities to get the public keys.

fn identities(&self) -> Identities

Get an iterator over the identities of this agent.

fn userauth(&self, username: &str, identity: &PublicKey) -> Result<(), Error>

Attempt public key authentication with the help of ssh-agent.