Struct ssh2::Channel

pub struct Channel<'a> {
    // some fields omitted
}

A channel represents a portion of an SSH connection on which data can be read and written.

Channels denote all of SCP uploads and downloads, shell sessions, remote process executions, and other general-purpose sessions.

Methods

impl<'a> Channel<'a>

unsafe fn from_raw(sess: &Session, raw: *mut LIBSSH2_CHANNEL) -> Channel

Wraps a raw pointer in a new Channel structure tied to the lifetime of the given session.

This consumes ownership of raw.

fn close(&mut self) -> Result<(), Error>

Close an active data channel.

In practice this means sending an SSH_MSG_CLOSE packet to the remote host which serves as instruction that no further data will be sent to it. The remote host may still send data back until it sends its own close message in response.

To wait for the remote end to close its connection as well, follow this command with wait_closed

fn wait_close(&mut self) -> Result<(), Error>

Enter a temporary blocking state until the remote host closes the named channel.

Typically sent after close in order to examine the exit status.

fn wait_eof(&mut self) -> Result<(), Error>

Wait for the remote end to acknowledge an EOF request.

fn eof(&self) -> bool

Check if the remote host has sent an EOF status for the selected stream.

fn process_startup(&mut self, request: &str, message: Option<&str>) -> Result<(), Error>

Initiate a request on a session type channel.

The SSH2 protocol currently defines shell, exec, and subsystem as standard process services.

fn request_pty(&mut self, term: &str, mode: Option<&str>, dim: Option<(uint, uint, uint, uint)>) -> Result<(), Error>

Request a PTY on an established channel.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

The dimensions argument is a tuple of (width, height, width_px, height_px)

fn request_pty_size(&mut self, width: uint, height: uint, width_px: Option<uint>, height_px: Option<uint>) -> Result<(), Error>

Request a PTY of a specified size

fn exec(&mut self, command: &str) -> Result<(), Error>

Execute a command

Example

let mut channel = session.channel_session().unwrap();
channel.exec("ls").unwrap();
println!("{}", channel.read_to_string().unwrap());

fn shell(&mut self) -> Result<(), Error>

Start a shell

fn subsystem(&mut self, system: &str) -> Result<(), Error>

Request a subsystem be started

fn flush_stream(&mut self, stream_id: uint) -> Result<(), Error>

Flush the read buffer for a given channel instance.

Groups of substreams may be flushed by passing on of the following constants

FLUSH_EXTENDED_DATA - Flush all extended data substreams

FLUSH_ALL - Flush all substreams

fn flush_stderr(&mut self) -> Result<(), Error>

Flush the stderr buffers.

fn write_stream(&mut self, stream_id: uint, data: &[u8]) -> Result<(), Error>

Write data to a channel stream.

All channel streams have one standard I/O substream (stream_id == 0), and may have up to 232 extended data streams as identified by the selected stream_id. The SSH2 protocol currently defines a stream ID of 1 to be the stderr substream.

fn write_stderr(&mut self, data: &[u8]) -> Result<(), Error>

Write data to the channel stderr stream.

fn exit_signal(&self) -> Result<ExitSignal, Error>

Get the remote exit signal.

fn exit_status(&self) -> Result<int, Error>

Returns the exit code raised by the process running on the remote host at the other end of the named channel.

Note that the exit status may not be available if the remote end has not yet set its status to closed.

fn read_stream(&mut self, stream_id: uint, data: &mut [u8]) -> Result<uint, Error>

Attempt to read data from an active channel stream.

All channel streams have one standard I/O substream (stream_id == 0), and may have up to 232 extended data streams as identified by the selected stream_id. The SSH2 protocol currently defines a stream ID of 1 to be the stderr substream.

fn read_stderr(&mut self, data: &mut [u8]) -> Result<uint, Error>

Read from the stderr stream .

fn setenv(&mut self, var: &str, val: &str) -> Result<(), Error>

Set an environment variable in the remote channel's process space.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

fn send_eof(&mut self) -> Result<(), Error>

Tell the remote host that no further data will be sent on the specified channel.

Processes typically interpret this as a closed stdin descriptor.

fn read_window(&self) -> ReadWindow

Check the status of the read window.

fn write_window(&self) -> WriteWindow

Check the status of the write window.

fn adjust_receive_window(&mut self, adjust: uint, force: bool) -> Result<uint, Error>

Adjust the receive window for a channel by adjustment bytes.

If the amount to be adjusted is less than the minimum adjustment and force is false, the adjustment amount will be queued for a later packet.

This function returns the new size of the receive window (as understood by remote end) on success.

fn limit_read(&mut self, limit: u64)

Artificially limit the number of bytes that will be read from this channel.