Struct ssh2::Channel

pub struct Channel<'sess> {
    // some fields omitted
}

A channel represents a portion of an SSH connection on which data can be read and written.

Channels denote all of SCP uploads and downloads, shell sessions, remote process executions, and other general-purpose sessions. Each channel implements the Reader and Writer traits to send and receive data. Whether or not I/O operations are blocking is mandated by the blocking flag on a channel's corresponding Session.

Methods

impl<'sess> Channel<'sess>

fn setenv(&mut self, var: &str, val: &str) -> Result<(), Error>

Set an environment variable in the remote channel's process space.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

fn request_pty(&mut self, term: &str, mode: Option<&str>, dim: Option<(u32, u32, u32, u32)>) -> Result<(), Error>

Request a PTY on an established channel.

Note that this does not make sense for all channel types and may be ignored by the server despite returning success.

The dimensions argument is a tuple of (width, height, width_px, height_px)

fn request_pty_size(&mut self, width: u32, height: u32, width_px: Option<u32>, height_px: Option<u32>) -> Result<(), Error>

Request a PTY of a specified size

fn exec(&mut self, command: &str) -> Result<(), Error>

Execute a command

An execution is one of the standard process services defined by the SSH2 protocol.

Example

let mut channel = session.channel_session().unwrap();
channel.exec("ls").unwrap();
let mut s = String::new();
channel.read_to_string(&mut s).unwrap();
println!("{}", s);

fn shell(&mut self) -> Result<(), Error>

Start a shell

A shell is one of the standard process services defined by the SSH2 protocol.

fn subsystem(&mut self, system: &str) -> Result<(), Error>

Request a subsystem be started.

A subsystem is one of the standard process services defined by the SSH2 protocol.

fn process_startup(&mut self, request: &str, message: Option<&str>) -> Result<(), Error>

Initiate a request on a session type channel.

The SSH2 protocol currently defines shell, exec, and subsystem as standard process services.

fn stderr<'a>(&'a mut self) -> Stream<'a, 'sess>

Get a handle to the stderr stream of this channel.

The returned handle implements the Read and Write traits.

fn stream<'a>(&'a mut self, stream_id: i32) -> Stream<'a, 'sess>

Get a handle to a particular stream for this channel.

The returned handle implements the Read and Write traits.

Groups of substreams may be flushed by passing on of the following constants and then calling flush().

FLUSH_EXTENDED_DATA - Flush all extended data substreams

FLUSH_ALL - Flush all substreams

fn exit_status(&self) -> Result<i32, Error>

Returns the exit code raised by the process running on the remote host at the other end of the named channel.

Note that the exit status may not be available if the remote end has not yet set its status to closed.

fn exit_signal(&self) -> Result<ExitSignal, Error>

Get the remote exit signal.

fn read_window(&self) -> ReadWindow

Check the status of the read window.

fn write_window(&self) -> WriteWindow

Check the status of the write window.

fn adjust_receive_window(&mut self, adjust: u64, force: bool) -> Result<u64, Error>

Adjust the receive window for a channel by adjustment bytes.

If the amount to be adjusted is less than the minimum adjustment and force is false, the adjustment amount will be queued for a later packet.

This function returns the new size of the receive window (as understood by remote end) on success.

fn limit_read(&mut self, limit: u64)

Artificially limit the number of bytes that will be read from this channel.

fn eof(&self) -> bool

Check if the remote host has sent an EOF status for the selected stream.

fn send_eof(&mut self) -> Result<(), Error>

Tell the remote host that no further data will be sent on the specified channel.

Processes typically interpret this as a closed stdin descriptor.

fn wait_eof(&mut self) -> Result<(), Error>

Wait for the remote end to acknowledge an EOF request.

fn close(&mut self) -> Result<(), Error>

Close an active data channel.

In practice this means sending an SSH_MSG_CLOSE packet to the remote host which serves as instruction that no further data will be sent to it. The remote host may still send data back until it sends its own close message in response.

To wait for the remote end to close its connection as well, follow this command with wait_closed

fn wait_close(&mut self) -> Result<(), Error>

Enter a temporary blocking state until the remote host closes the named channel.

Typically sent after close in order to examine the exit status.