Struct ssh2::Error

pub struct Error {
    // some fields omitted
}

Representation of an error that can occur within libssh2

Methods

impl Error

fn last_error(sess: &Session) -> Option<Error>

Generate the last error that occurred for a Session.

Returns None if there was no last error.

fn new(code: c_int, msg: &'static str) -> Error

Create a new error for the given code and message

fn eof() -> Error

Generate an error that represents EOF

fn from_errno(code: c_int) -> Error

Construct an error from an error code from libssh2

fn message(&self) -> &str

Get the message corresponding to this error

fn code(&self) -> c_int

Return the code for this error