Struct ssh2::ExitSignal

pub struct ExitSignal {
    pub exit_signal: Option<String>,
    pub error_message: Option<String>,
    pub lang_tag: Option<String>,
}

Data received from when a program exits with a signal.

Fields

exit_signal

The exit signal received, if the program did not exit cleanly. Does not contain a SIG prefix

error_message

Error message provided by the remote server (if any)

lang_tag

Language tag provided by the remote server (if any)