Struct ssh2::File

pub struct File<'sftp> {
    // some fields omitted
}

A file handle to an SFTP connection.

Files behave similarly to std::old_io::File in that they are readable and writable and support operations like stat and seek.

Files are created through open, create, and open_mode on an instance of Sftp.

Methods

impl<'sftp> File<'sftp>

fn setstat(&mut self, stat: FileStat) -> Result<(), Error>

Set the metadata for this handle.

fn stat(&mut self) -> Result<FileStat, Error>

Get the metadata for this handle.

fn statvfs(&mut self) -> Result<LIBSSH2_SFTP_STATVFS, Error>

fn readdir(&mut self) -> Result<(PathBuf, FileStat), Error>

Reads a block of data from a handle and returns file entry information for the next entry, if any.

Note that this provides raw access to the readdir function from libssh2. This will return an error when there are no more files to read, and files such as . and .. will be included in the return values.

Also note that the return paths will not be absolute paths, they are the filenames of the files in this directory.

fn fsync(&mut self) -> Result<(), Error>

This function causes the remote server to synchronize the file data and metadata to disk (like fsync(2)).

For this to work requires fsync@openssh.com support on the server.