Struct ssh2::Host

pub struct Host<'kh> {
    // some fields omitted
}

Structure representing a known host as part of a KnownHosts structure.

Methods

impl<'kh> Host<'kh>

fn name(&self) -> Option<&str>

This is None if no plain text host name exists.

fn key(&self) -> &str

Returns the key in base64/printable format