Struct ssh2::Host

pub struct Host<'kh> {
    // some fields omitted
}

Structure representing a known host as part of a KnownHosts structure.

Methods

impl<'kh> Host<'kh>

unsafe fn from_raw(raw: *mut libssh2_knownhost) -> Host<'kh>

Creates a new host from its raw counterpart.

Unsafe as there are no restrictions on the lifetime of the host returned and the validity of raw is not known.

fn name(&self) -> Option<&str>

This is None if no plain text host name exists.

fn key(&self) -> &str

Returns the key in base64/printable format

fn raw(&self) -> *mut libssh2_knownhost

Gain access to the underlying raw pointer