Struct ssh2::Listener

pub struct Listener<'sess> {
    // some fields omitted
}

A listener represents a forwarding port from the remote server.

New channels can be accepted from a listener which represent connections on the remote server's port.

Methods

impl<'sess> Listener<'sess>

fn accept(&mut self) -> Result<Channel<'sess>, Error>

Accept a queued connection from this listener.