Struct ssh2::ReadWindow

pub struct ReadWindow {
    pub remaining: u32,
    pub available: u32,
    pub window_size_initial: u32,
}

Description of the read window as returned by Channel::read_window

Fields

remaining

The number of bytes which the remote end may send without overflowing the window limit.

available

The number of bytes actually available to be read.

window_size_initial

The window_size_initial as defined by the channel open request