Struct ssh2::Session

pub struct Session {
    // some fields omitted
}

An SSH session, typically representing one TCP connection.

All other structures are based on an SSH session and cannot outlive a session. Sessions are created and then have the TCP socket handed to them (via the handshake method).

Methods

impl Session

fn new() -> Option<Session>

Initializes an SSH session object.

unsafe fn from_raw(raw: *mut LIBSSH2_SESSION) -> Session

Takes ownership of the given raw pointer and wraps it in a session.

This is unsafe as there is no guarantee about the validity of raw.

fn banner(&self) -> Option<&str>

Get the remote banner

Once the session has been setup and handshake() has completed successfully, this function can be used to get the server id from the banner each server presents.

May return None on invalid utf-8 or if an error has ocurred.

fn banner_bytes(&self) -> Option<&[u8]>

See banner.

Will only return None if an error has ocurred.

fn set_banner(&self, banner: &str) -> Result<(), Error>

Set the SSH protocol banner for the local client

Set the banner that will be sent to the remote host when the SSH session is started with handshake(). This is optional; a banner corresponding to the protocol and libssh2 version will be sent by default.

fn disconnect(&self, reason: Option<DisconnectCode>, description: &str, lang: Option<&str>) -> Result<(), Error>

Terminate the transport layer.

Send a disconnect message to the remote host associated with session, along with a reason symbol and a verbose description.

fn flag(&self, flag: SessionFlag, enable: bool) -> Result<(), Error>

Enable or disable a flag for this session.

fn is_blocking(&self) -> bool

Returns whether the session was previously set to nonblocking.

fn set_blocking(&self, blocking: bool)

Set or clear blocking mode on session

Set or clear blocking mode on the selected on the session. This will instantly affect any channels associated with this session. If a read is performed on a session with no data currently available, a blocking session will wait for data to arrive and return what it receives. A non-blocking session will return immediately with an empty buffer. If a write is performed on a session with no room for more data, a blocking session will wait for room. A non-blocking session will return immediately without writing anything.

fn timeout(&self) -> uint

Returns the timeout, in milliseconds, for how long blocking calls may wait until they time out.

A timeout of 0 signifies no timeout.

fn set_timeout(&self, timeout_ms: uint)

Set timeout for blocking functions.

Set the timeout in milliseconds for how long a blocking the libssh2 function calls may wait until they consider the situation an error and return an error.

By default or if you set the timeout to zero, libssh2 has no timeout for blocking functions.

fn host_key(&self) -> Option<(&[u8], HostKeyType)>

Get the remote key.

Returns None if something went wrong.

fn host_key_hash(&self, hash: HashType) -> Option<&[u8]>

Returns the computed digest of the remote system's hostkey.

The bytes returned are the raw hash, and are not printable. If the hash is not yet available None is returned.

fn method_pref(&self, method_type: MethodType, prefs: &str) -> Result<(), Error>

Set preferred key exchange method

The preferences provided are a comma delimited list of preferred methods to use with the most preferred listed first and the least preferred listed last. If a method is listed which is not supported by libssh2 it will be ignored and not sent to the remote host during protocol negotiation.

fn methods(&self, method_type: MethodType) -> Option<&str>

Return the currently active algorithms.

Returns the actual method negotiated for a particular transport parameter. May return None if the session has not yet been started.

fn supported_algs(&self, method_type: MethodType) -> Result<Vec<&'static str>, Error>

Get list of supported algorithms.

fn agent(&self) -> Result<Agent, Error>

Init an ssh-agent handle.

The returned agent will still need to be connected manually before use.

fn handshake(&mut self, socket: libssh2_socket_t) -> Result<(), Error>

Begin transport layer protocol negotiation with the connected host.

The socket provided is a connected socket descriptor. Typically a TCP connection though the protocol allows for any reliable transport and the library will attempt to use any berkeley socket.

fn channel_open(&self, channel_type: &str, window_size: uint, packet_size: uint, message: Option<&str>) -> Result<Channel, Error>

Allocate a new channel for exchanging data with the server.

This is typically not called directly but rather through channel_open_session, channel_direct_tcpip, or channel_forward_listen.

fn channel_session(&self) -> Result<Channel, Error>

Establish a new session-based channel.

fn channel_direct_tcpip(&self, host: &str, port: u16, src: Option<(&str, u16)>) -> Result<Channel, Error>

Tunnel a TCP connection through an SSH session.

Tunnel a TCP/IP connection through the SSH transport via the remote host to a third party. Communication from the client to the SSH server remains encrypted, communication from the server to the 3rd party host travels in cleartext.

The optional src argument is the host/port to tell the SSH server where the connection originated from.

fn channel_forward_listen(&self, remote_port: u16, host: Option<&str>, queue_maxsize: Option<uint>) -> Result<(Listener, u16), Error>

Instruct the remote SSH server to begin listening for inbound TCP/IP connections.

New connections will be queued by the library until accepted by forward_accept.

fn userauth_password(&self, username: &str, password: &str) -> Result<(), Error>

Attempt basic password authentication.

Note that many SSH servers which appear to support ordinary password authentication actually have it disabled and use Keyboard Interactive authentication (routed via PAM or another authentication backed) instead.

fn userauth_pubkey_file(&self, username: &str, pubkey: Option<&Path>, privatekey: &Path, passphrase: Option<&str>) -> Result<(), Error>

Attempt public key authentication using a PEM encoded private key file stored on disk.

fn userauth_hostbased_file(&self, username: &str, publickey: &Path, privatekey: &Path, passphrase: Option<&str>, hostname: &str, local_username: Option<&str>) -> Result<(), Error>

Umm... I wish this were documented in libssh2?

fn authenticated(&self) -> bool

Indicates whether or not the named session has been successfully authenticated.

fn auth_methods(&self, username: &str) -> Result<&str, Error>

Send a SSH_USERAUTH_NONE request to the remote host.

Unless the remote host is configured to accept none as a viable authentication scheme (unlikely), it will return SSH_USERAUTH_FAILURE along with a listing of what authentication schemes it does support. In the unlikely event that none authentication succeeds, this method with return NULL. This case may be distinguished from a failing case by examining libssh2_userauth_authenticated.

The return value is a comma-separated string of supported auth schemes.

fn keepalive_set(&self, want_reply: bool, interval: uint) -> Result<(), Error>

Set how often keepalive messages should be sent.

The want_reply argument indicates whether the keepalive messages should request a response from the server.

The interval argument is number of seconds that can pass without any I/O, use 0 (the default) to disable keepalives. To avoid some busy-loop corner-cases, if you specify an interval of 1 it will be treated as 2.

fn keepalive_send(&self) -> Result<uint, Error>

Send a keepalive message if needed.

Returns how many seconds you can sleep after this call before you need to call it again.

fn known_hosts(&self) -> Result<KnownHosts, Error>

Init a collection of known hosts for this session.

Returns the handle to an internal representation of a known host collection.

fn scp_recv(&self, path: &Path) -> Result<(Channel, FileStat), Error>

Request a file from the remote host via SCP.

fn scp_send(&self, remote_path: &Path, mode: FilePermission, size: u64, times: Option<(u64, u64)>) -> Result<Channel, Error>

Send a file to the remote host via SCP.

The remote_path provided will the remote file name. The times argument is a tuple of (mtime, atime), and will default to the remote host's current time if not specified.

fn sftp(&self) -> Result<Sftp, Error>

Open a channel and initialize the SFTP subsystem.

Although the SFTP subsystem operates over the same type of channel as those exported by the Channel API, the protocol itself implements its own unique binary packet protocol which must be managed with the methods on Sftp.

fn raw(&self) -> *mut LIBSSH2_SESSION

Gain access to the underlying raw libssh2 session pointer.

fn rc(&self, rc: c_int) -> Result<(), Error>

Translate a return code into a Rust-Result.