Struct ssh2::Sftp

pub struct Sftp<'sess> {
    // some fields omitted
}

A handle to a remote filesystem over SFTP.

Instances are created through the sftp method on a Session.

Methods

impl<'sess> Sftp<'sess>

fn open_mode(&self, filename: &Path, flags: OpenFlags, mode: i32, open_type: OpenType) -> Result<File, Error>

Open a handle to a file.

fn open(&self, filename: &Path) -> Result<File, Error>

Helper to open a file in the Read mode.

fn create(&self, filename: &Path) -> Result<File, Error>

Helper to create a file in write-only mode with truncation.

fn opendir(&self, dirname: &Path) -> Result<File, Error>

Helper to open a directory for reading its contents.

fn readdir(&self, dirname: &Path) -> Result<Vec<(PathBuf, FileStat)>, Error>

Convenience function to read the files in a directory.

The returned paths are all joined with dirname when returned, and the paths . and .. are filtered out of the returned list.

fn mkdir(&self, filename: &Path, mode: i32) -> Result<(), Error>

Create a directory on the remote file system.

fn rmdir(&self, filename: &Path) -> Result<(), Error>

Remove a directory from the remote file system.

fn stat(&self, filename: &Path) -> Result<FileStat, Error>

Get the metadata for a file, performed by stat(2)

fn lstat(&self, filename: &Path) -> Result<FileStat, Error>

Get the metadata for a file, performed by lstat(2)

fn setstat(&self, filename: &Path, stat: FileStat) -> Result<(), Error>

Set the metadata for a file.

fn symlink(&self, path: &Path, target: &Path) -> Result<(), Error>

Create a symlink at target pointing at path.

fn readlink(&self, path: &Path) -> Result<PathBuf, Error>

Read a symlink at path.

fn realpath(&self, path: &Path) -> Result<PathBuf, Error>

Resolve the real path for path.

fn rename(&self, src: &Path, dst: &Path, flags: Option<RenameFlags>) -> Result<(), Error>

Rename a filesystem object on the remote filesystem.

The semantics of this command typically include the ability to move a filesystem object between folders and/or filesystem mounts. If the Overwrite flag is not set and the destfile entry already exists, the operation will fail.

Use of the other flags (Native or Atomic) indicate a preference (but not a requirement) for the remote end to perform an atomic rename operation and/or using native system calls when possible.

If no flags are specified then all flags are used.

fn unlink(&self, file: &Path) -> Result<(), Error>

Remove a file on the remote filesystem

fn last_error(&self) -> Error

Peel off the last error to happen on this SFTP instance.

fn rc(&self, rc: c_int) -> Result<(), Error>

Translates a return code into a Rust-Result