Struct ssh2::WriteWindow

pub struct WriteWindow {
    pub remaining: uint,
    pub window_size_initial: uint,
}

Description of the write window as returned by Channel::write_window

Fields

remaining

The number of bytes which may be safely written on the channel without blocking.

window_size_initial

The window_size_initial as defined by the channel open request